+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.0 (2018-06-02)
+
++ Allow fields in variant
++ Add attribute: `value`
+
 # 0.1.0 (2017-09-14)
 
 ## Added